refactor: #20 Todo 렌더링 리팩토링
