+++ src/components/TodoItem/types.ts
@@ -1,8 +1,6 @@
 import { TodoType } from '../TodoList/types'
 
-interface TodoItemProps extends Omit<TodoType, 'userId'> {
+export interface TodoItemProps extends Omit<TodoType, 'userId'> {
   updateTodo: (id: number, todo: string, isCompleted: boolean) => void
   deleteTodo: (id: number) => void
 }
-
-export default TodoItemProps
